Add complete search functionality for samples
