+++ src/py/aspen/cli/db.py
@@ -3,7 +3,7 @@ from IPython.terminal.embed import InteractiveShellEmbed
 
 from aspen.cli.toplevel import cli
 from aspen.config.development import DevelopmentConfig
-from aspen.covidhub_import import import_project
+# from aspen.covidhub_import.implementation import import_project
 from aspen.database.connection import get_db_uri, init_db
 from aspen.database.models import *  # noqa: F401, F403
 from aspen.database.schema import create_tables_and_schema
@@ -32,12 +32,12 @@ def interact(ctx):
     shell()
 
 
-@db.command("import-covidhub-project")
-@click.option("--covidhub-db-secret", default="cliahub/cliahub_test_db")
-@click.option("--rr-project-id", type=str, required=True)
-@click.pass_context
-def import_covidhub_project(ctx, covidhub_db_secret, rr_project_id):
-    # these are injected into the IPython scope, but they appear to be unused.
-    engine = ctx.obj["ENGINE"]
+# @db.command("import-covidhub-project")
+# @click.option("--covidhub-db-secret", default="cliahub/cliahub_test_db")
+# @click.option("--rr-project-id", type=str, required=True)
+# @click.pass_context
+# def import_covidhub_project(ctx, covidhub_db_secret, rr_project_id):
+#     # these are injected into the IPython scope, but they appear to be unused.
+#     engine = ctx.obj["ENGINE"]
 
-    import_project(engine, covidhub_db_secret, rr_project_id)
+#     import_project(engine, covidhub_db_secret, rr_project_id)
