+++ src/ts/package.json
@@ -33,7 +33,6 @@
     "case-sensitive-paths-webpack-plugin": "2.3.0",
     "classnames": "^2.2.6",
     "css-loader": "4.3.0",
-    "cz-ui": "^0.4.18",
     "dotenv": "8.2.0",
     "dotenv-expand": "5.1.0",
     "eslint": "^7.11.0",
@@ -51,7 +50,7 @@
     "fs-extra": "^9.0.1",
     "html-webpack-plugin": "4.5.0",
     "identity-obj-proxy": "3.0.0",
-    "lodash": "^4.17.20",
+    "lodash": "^4.17.21",
     "mini-css-extract-plugin": "0.11.3",
     "node-sass": "^5.0.0",
     "optimize-css-assets-webpack-plugin": "5.0.4",
