--- src/ts/src/common/types/cz-ui.d.ts
@@ -1,5 +1,0 @@
-/// <reference types="node" />
-/// <reference types="react" />
-/// <reference types="react-dom" />
-
-declare module "cz-ui";
